fix: Remove extra tab stop on XML group (#82)
Ben's PR #72 removed the tabindex on the enclosing DIV, but there's
also a tabindex on the XML element that encloses the G that gets
the focus/blur event listeners.  Remove it.

An alternative approach would be to add the event listeners to the
XML element instead of the G element, and this might be preferable
overall but breaks the visual indicator that you have tabbed to the
workspace (and I don't know how to fix it) so I'm going with this
approach for now.