feat: Simplify workspace navigation flow (#126)
Fixes part of #101
Fixes #89

This change removes the need for using enter and escape to enable keyboard navigation when the workspace is focused.

Instead, simply focusing the workspace is sufficient to allow keyboard navigation. Note that this PR goes back to the previous behavior of always enabling keyboard navigation on page load and instead limits the preconditions of navigation keys to check focus rather than feature enabled state. This allows for #89 to be fully mitigated. Since keyboard navigation is never disabled, the cursor position is never lost (and thus doesn't need to be cached). The removal of enter/escape fixes the first part of #101.

This PR also performs some other minor cleanup work in ``navigation_controller.ts``.