fix: Move cursor when user selects a block with mouse
Hack to temporarily address #97.

Overrides Marker.prototype.getCurNode in LineCursor to add a
hack that attempts to detect if the user has moved focus by
selecting a block and, if so, update the cursor location (and
any highlighting) to match.

This works reasonably well but has some glitches, most notably
that if the cursor is not on a block (e.g. it is on a connection
or the workspace) then it will remain visible in its previous
location until a cursor key is pressed.