feat: show some block for context when opening action menu (#191)
Previously it was shown over the top left, which obscured useful context.  I
considered positioning it below the block but it can be a long way down in some
cases. That can feel unexpected and makes it more likely that the menu is
repositioned to keep it on screen.

Mentioned on https://github.com/google/blockly-keyboard-experimentation/issues/184#issuecomment-2623882452

This is a less arbitrary version of the offset I had in "Demo A" (that just
used 50px which looked fine for zelos styling but not great in geras).