feat: Plumb CursorOptions through KeyboardNavigation
Add a NavigationOptions type and use it for an options object
for the KeyboardNavigation constructor.  For now it just has
a single property of type CursorOptions, but I anticipate
adding flags for different test scenarios (e.g. which key opens
context menu, on which nodes various options are presented).

Will plumb this through the test code in the next RP.