feat: Add options for renderer and stackConnections
Introduce getOptions function (replacing most of loadScenario)
to parse query parameters for the scenario, renderer and
stackConnection options and update the form controls to match
(after a short delay so that Chrome does not later restore
incorrect form state when page navigation has occurred).

N.B.: Because checkbox inputs do not send any parameter when
not checked, it is not possible (without considerable extra
trouble) to have a checkbox default to checked, so the
checkbox for stackConnections is inverted from the option
and is labelled accordingly.

Fixes #179.