fix: Put cursor on newly-created block, not its top connection
Set the cursor's curNode to a BLOCK node rather than a PREVIOUS
or OUTPUT node.

Fixes #221.

Note that due to issues with how Blockly renders markers,
this change results in the block being both selected and
having a block marker, though the latter disappears as soon
as the cursor is moved.