fix: Prevent renderer from making block marker SVG reappear
Override Marker.prototype.setCurNode and .draw to prevent those
methods from calling this.drawer.draw; instead, route those
calls via drawMarker (a revised version of the previous
updateFocusIndication method).

This ensures that this.drawer.draw() is only called if curNode
is not a BLOCK node, even when the block to which the marker is
attached is being rendered.

Since MARKER_MOVED events are normally fired by
MarkerSvg.prototype.draw calling this.fireMarkerEvent, drawMarker
will arrange to call .fireMarkerEvent directly when it does not
call this.drawer.draw.