fix: Put cursor on newly-created block, not its top connection (#252)
* fix: Put cursor on newly-created block, not its top connection

Set the cursor's curNode to a BLOCK node rather than a PREVIOUS
or OUTPUT node.

Fixes #221.

Note that due to issues with how Blockly renders markers,
this change results in the block being both selected and
having a block marker, though the latter disappears as soon
as the cursor is moved.

* fix: Prevent renderer from making block marker SVG reappear

Override Marker.prototype.setCurNode and .draw to prevent those
methods from calling this.drawer.draw; instead, route those
calls via drawMarker (a revised version of the previous
updateFocusIndication method).

This ensures that this.drawer.draw() is only called if curNode
is not a BLOCK node, even when the block to which the marker is
attached is being rendered.

Since MARKER_MOVED events are normally fired by
MarkerSvg.prototype.draw calling this.fireMarkerEvent, drawMarker
will arrange to call .fireMarkerEvent directly when it does not
call this.drawer.draw.