feat: Only show Edit option when useful
If there are no inline nodes to the right of the cursor on the
'current line' of the program, hide the Edit option.

This means the menu item is shown any time the cursor is on a
block and not in the rightmost position 'on the current line';
consequently sometimes the description ("Edit block contents")
is possibly misleading, because it might not be the contents of
the _current_ block that's being navigated to, but rather that
of a sibling or parent block.