feat: Make toolbox & flyout properly focusable. (#225)
* feat: Make toolbox & flyout properly focusable.

The toolbox already had a tab stop, but it wouldn't become properly
enabled for keyboard navigation. Now it is properly enabled and works
when focused both via tab navigation and with the 'T' shortcut.

* chore: address reviewer comments

* feat: add subcategory to demo toolbox categories

This helps to verify subcategory behavior (which has been observed as
not completely working currently).

* Address reviewer comment for monkey patch install.

* Revert monkeypatch installation changes.

* Add fix for null current node.

It seems possible now to call hide() earlier than previously expected
(i.e. when there's a cursor but no current node, such as in the case of
focusing the toolbox before the workspace as this PR now makes
possible).