feat: Edit option in block context menu (#274)
* refactor: Rename shortcuts to match arrow keys

Rename the main directional shortcuts to match the names of
corresponding keys.

Not renaming the corresponding methods on LineCursor yet
as that seems like a riskier change.

* chore: Sort inputs in navigation_controller.ts

* feat: Edit action

* feat: Only show Edit option when useful

If there are no inline nodes to the right of the cursor on the
'current line' of the program, hide the Edit option.

This means the menu item is shown any time the cursor is on a
block and not in the rightmost position 'on the current line';
consequently sometimes the description ("Edit block contents")
is possibly misleading, because it might not be the contents of
the _current_ block that's being navigated to, but rather that
of a sibling or parent block.

* fix: Adjust weights to put edit between insert and delete

* fix: typo in src/constants.ts

Co-authored-by: Matt Hillsdon <matt.hillsdon@microbit.org>

* fix: Revert f05269f for PR #274.

* fix: JSDoc tweaks for PR #274

* fix: remove dead import

---------

Co-authored-by: Matt Hillsdon <matt.hillsdon@microbit.org>
Co-authored-by: Rachel Fenichel <fenichel@google.com>