feat: consistent focus handling; remove markedNode
Ensure we always focus the same element when keyboard navigation works in that
element. Previously the toolbox sometimes was and sometimes wasn't focused when
it navigated, the workspace or group or the workspace SVG could be focused and
the flyout never took the focus.

With this approach the relevant container takes focus.

Manage the passive focus indicator based on focus rather than markedNode state.

Integrate with Gesture to determine how mouse behaviours relate to focus. This
prevents issues where a gesture would cause a focus change which would close
the flyout, interrupting a normal drag. In time hopefully this code can move to
core where it will be more natural and less fragile.

Fixes #227
Fixes #200
Part of #229 (scrollbars still an issue)
Fixes #222