fix: take care making a selection when focusing the workspace
This can interfere with in-progress gestures causing them to act in unexpected
ways. We now clear the current node when focus moves away.

Avoid altering the selection in cursor.draw to prevent similar poorly timed
selections.

If the cursor position is lost, e.g. by click in the workspace, the arrow keys
now default it if needed. Tabbing to the workspace always shows a cursor.

Fixes:
- Click in workspace blank space now doesn't select a block on mouse down
- The mini-workspace for "if" and similar now behaves correctly again (no
  keyboard nav still though)
- Also fixes #287