feat: consistent focus handling; remove markedNode (#326)
* feat: consistent focus handling; remove markedNode

Ensure we always focus the same element when keyboard navigation works in that
element. Previously the toolbox sometimes was and sometimes wasn't focused when
it navigated, the workspace or group or the workspace SVG could be focused and
the flyout never took the focus.

With this approach the relevant container takes focus.

Manage the passive focus indicator based on focus rather than markedNode state.

Integrate with Gesture to determine how mouse behaviours relate to focus. This
prevents issues where a gesture would cause a focus change which would close
the flyout, interrupting a normal drag. In time hopefully this code can move to
core where it will be more natural and less fragile.

Fixes #227
Fixes #200
Part of #229 (scrollbars still an issue)
Fixes #222

* fix: take care making a selection when focusing the workspace

This can interfere with in-progress gestures causing them to act in unexpected
ways. We now clear the current node when focus moves away.

Avoid altering the selection in cursor.draw to prevent similar poorly timed
selections.

If the cursor position is lost, e.g. by click in the workspace, the arrow keys
now default it if needed. Tabbing to the workspace always shows a cursor.

Fixes:
- Click in workspace blank space now doesn't select a block on mouse down
- The mini-workspace for "if" and similar now behaves correctly again (no
  keyboard nav still though)
- Also fixes #287

* fix: cope with null node

* fix: remove unintended clearing of current node

* fix: flyout-only paste fix

* fix: selection in paste in flyout case

* chore: clarify case this applies