fix: right click menu issues
PR #326 introduced an issue with stale current node when showing the context
menu by mouse. It's only possible to fix this by reinstating the updating of
curNode from the selection.

We also attempt to avoid showing passive focus when using the context menu and
widgets (via mouse and keyboard) for consistency with the normal Blockly
experience.

Fixes an issue clicking on shadow blocks which could previously show an outline
around the shadow and parent block.