fix: account for ephemeral focus in enter precondition  (#619)
* chore: upgrade field colour to buggy version

This illustrates that the latest version, based on field-grid-dropdown, does
not work with keyboard navigation enabled. This is because Enter propagates
to the global handler, causing the field editor to be retriggered or block
toast messages to be shown.

* test: inverted test for colour field enter

Currently with 6.0.2 this triggers the global shortcuts.

* fix: account for ephemeral focus in enter precondition

Previously bugs could occur if the field editor contained a button but did
prevent keydown events propagating.