fix: prevent enter on icons moving to next icon/field (#638)
The in() behaviour relied on here only works for MutatorIcons. See
icon_navigation_policy.ts in core.

That might change over time, but there will still be some icons for which this
isn't appropriate.

Prior to this change enter on a comment icon or a MakeCode debug breakpoint
icon inexplicably moves the cursor to the next field/icon.