chore: lint for floating/misused promises (#634)
Fix missing awaits in the tests.

Needed tsconfig information so I've added the project setting - see
https://typescript-eslint.io/blog/parser-options-project-true/#introducing-true

I've done this just for the tests as otherwise we hit issues with the test app
and the webdriver test app not being covered by tsconfig.json files. It could
be trivially extended to the src tree.