feat: Add initial screen reader support.
This is in the form of a bunch of monkey patches that carefully replace
essentially all ARIA initialization happening within core Blockly.

This is largely done in a way that won't be well compatible with certain
uses of Blockly itself, and is only mean to act as a baseline for
testing.