feat: Deploy GH pages with screen reader experiment (#689)
Fixes #686

This updates the GH pages workflow to:
- Automatically deploy with changes to both `main` and `add-screen-reader-support-experimental` (the screen reader experimental branch--see https://github.com/google/blockly/issues/9283).
- Upload the screen reader version of the playground to a new `/screenreader` subdirectory in the plugin repo's GitHub pages deployment.

This will approximately double the amount of time needed for the GH pages deployment since it requires building two distinct copies of core Blockly and the plugin (tip-of-tree versions and the screen reader experimentation versions).

Conveniently, this will act as a fast way to see screen reader changes since they will go live a few minutes after a change to the plugin. Unfortunately, that does not reciprocate for changes to core Blockly (where most of the screen reader changes will actually occur), so oftentimes a manual run of the workflow will be needed (which should suffice in those cases).

Note that this needs to be approached in this way (building and uploading both projects for changes to either branch) since the action being used for deployment doesn't support partial deployment. See: https://github.com/actions/deploy-pages/issues/349.