test(Mover): Also check index of candidateConnection.local
Modify getConnectionCandidate to also return the index of
candidateConnection.local in the list of connections returned
from block.getConnections_(true), so we can check wich connection
on the moving block is being used to connect.

For the simple mover case this is not very interesting (it's
normally index 0, for the previous connection, occasionally
index 1 if the moving block will be the new top block in a stack)
but this will be more important for tests involving moving more
complexly-shaped blocks.