test(Mover): Add value expression right/left move tests
Moving value blocks left/right behaves as expected when using
geras, but in zelos there is a bug where external inputs
are not visited in the expected order (not that external inputs
_look_ external when using geras, but still it is a possible
configuration and should behave as expected).

Note that two of these tests fail due to a bug; this will be dealt
with in the next commit.